refactor(shortint): move scalar_div_mod to its own modulue
